BAU: Refactor - move EVCS code into evcs namespace
