+++ libs/evcs-service/build.gradle
@@ -49,7 +49,7 @@ test {
 
 task pactConsumerTests (type: Test) {
 	useJUnitPlatform()
-	include 'uk/gov/di/ipv/core/library/pact/**'
+	include 'uk/gov/di/ipv/core/library/evcs/pact/**'
 	systemProperties['pact.rootDir'] = "$rootDir/build/pacts"
 }
