+++ libs/evcs-service/src/main/java/uk/gov/di/ipv/core/library/evcs/dto/EvcsGetUserVCsDto.java
@@ -1,4 +1,4 @@
-package uk.gov.di.ipv.core.library.dto;
+package uk.gov.di.ipv.core.library.evcs.dto;
 
 import java.util.List;
