+++ libs/evcs-service/src/main/java/uk/gov/di/ipv/core/library/evcs/dto/EvcsUpdateUserVCsDto.java
@@ -1,7 +1,7 @@
-package uk.gov.di.ipv.core.library.dto;
+package uk.gov.di.ipv.core.library.evcs.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import uk.gov.di.ipv.core.library.enums.EvcsVCState;
+import uk.gov.di.ipv.core.library.evcs.enums.EvcsVCState;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public record EvcsUpdateUserVCsDto(String signature, EvcsVCState state, Object metadata) {}
