+++ libs/evcs-service/src/main/java/uk/gov/di/ipv/core/library/evcs/enums/EvcsVCState.java
@@ -1,4 +1,4 @@
-package uk.gov.di.ipv.core.library.enums;
+package uk.gov.di.ipv.core.library.evcs.enums;
 
 import uk.gov.di.ipv.core.library.annotations.ExcludeFromGeneratedCoverageReport;
