+++ libs/evcs-service/src/main/java/uk/gov/di/ipv/core/library/evcs/metadata/InheritedIdentityMetadata.java
@@ -1,3 +1,3 @@
-package uk.gov.di.ipv.core.library.metadata;
+package uk.gov.di.ipv.core.library.evcs.metadata;
 
 public record InheritedIdentityMetadata(String inheritedIdentity) {}
